chore: centralize app layout and update links
