--- layout.tsx
@@ -1,47 +1,0 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        <div className="flex h-screen">
-          <aside className="w-64 bg-gray-800 text-white p-4">
-            <nav>
-              <ul>
-                <li><a href="/dashboard">Dashboard</a></li>
-                <li><a href="/contacts">Contacts</a></li>
-                <li><a href="/settings">Settings</a></li>
-              </ul>
-            </nav>
-          </aside>
-          <main className="flex-1 p-4">
-            {children}
-          </main>
-        </div>
-      </body>
-    </html>
-  );
-}
